requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug==2.3.7
flask-login==0.6.2
flask-bcrypt==1.0.1
schedule==1.2.0
pillow==10.0.1
twilio==8.10.0  # For SMS
flask-mail==0.9.1  # For Email
apscheduler==3.10.4